Move Schema from mentat_db to mentat_core.
Also define SchemaMap in terms of Entid, not i64.
